Adding delete and update test result

+++ routes/test.js
@@ -25,7 +25,7 @@ app.post("/test/view", verifyJWT, (req, res) => {
   const userID = req.userID;
   const { testID } = req.body;
 
-  Test.findOne({ testID: testID }, (err, userTestResult) => {
+  Test.findOne({ testID: testID, userID: userID }, (err, userTestResult) => {
     if (err) {
       res.json({
         success: false,
@@ -68,11 +68,12 @@ app.post("/test/new", verifyJWT, (req, res) => {
 //Edit a previous test or delete test
 app.post("/test/modify", verifyJWT, (req, res) => {
   //Action denotes the action to perform: delete, update test verdict;
+  const userID = req.userID;
   const { testID, action, content } = req.body;
   switch (action) {
     case "delete":
       //Delete user test
-      Test.deleteOne({ testID: `${testID}` })
+      Test.deleteOne({ testID: `${testID}`, userID: userID })
         .then((queryResponse) => {
           console.log("Data deleted: "), queryResponse;
           res.json({
@@ -89,7 +90,7 @@ app.post("/test/modify", verifyJWT, (req, res) => {
     case "update":
       //Update user verdict for test
       //Content from req.body will contain user verdict
-      const query = { testID: testID };
+      const query = { testID: testID, userID: userID };
       Test.findOneAndUpdate(
         query,
         { userVerdict: content.userVerdict },
@@ -99,12 +100,7 @@ app.post("/test/modify", verifyJWT, (req, res) => {
             success: true,
           });
         }
-      ).catch((err) => {
-        console.error(err);
-        res.json({
-          success: false,
-        });
-      });
+      );
       break;
   }
 });
